requirements.txt
pytest==9.0.2
selene==2.0.0rc9
selenium==4.39.0
allure-pytest>=2.15.0
allure-python-commons